refactor(blog): extract schema logic to useBlogDetailSchema composable
- Moved schema definition logic from [slug].vue into a dedicated composable for better modularity.
- Simplified the blog page component by delegating schema generation to the new composable.
